% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_BBMV_circular_trait.r
\name{DiffMat_backwards_circular_trait}
\alias{DiffMat_backwards_circular_trait}
\title{Npts+1'=0 on the circle: it is where it binds}
\usage{
DiffMat_backwards_circular_trait(V)
}
\description{
Npts+1'=0 on the circle: it is where it binds
}
