% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitWrappedBM.R
\name{fitWrappedBM}
\alias{fitWrappedBM}
\title{Model fitting for continuous comparative data in circular space}
\usage{
fitWrappedBM(phy, trait, bounds)
}
\arguments{
\item{phy}{a phylogenetic tree of class phylo}

\item{trait}{data vector for a single trait, with names matching tips in phy}

\item{bounds}{range defining bounds of circular space}
}
\description{
Fitting wrapped Brownian Motion model to phylogenetic trees
}
\examples{
tree = ape::rcoal(10)
trait = runif(n = 10, min = 0, max = 1)
names(trait) = tree$tip.label
fitWrappedBM(phy = tree, trait = trait, bounds = c(0, 1))
}
\author{
Florian Boucher and Mark Juhn
}
