% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedPIC.R
\name{wrappedPIC}
\alias{wrappedPIC}
\title{Wrapped Phylogenetically Independent Contrasts}
\usage{
wrappedPIC(trait, phy, lower_bound, upper_bound, return_ASE = F)
}
\arguments{
\item{trait}{data vector for a single trait, with names matching tips in phy}

\item{phy}{a phylogenetic tree of class phylo}

\item{lower_bound}{lower bound for trait in circular space}

\item{upper_bound}{upper bound for trait in circular space}

\item{return_ASE}{logical, if TRUE, the tip values and ancestral state estimations are returned along with the wrapped phylogenetically independent contrasts}
}
\description{
Computes phylogenetically independent contrasts in circular space
}
\examples{
tree = ape::rcoal(10)
trait = runif(n = 10, min = 0, max = 1)
names(trait) = tree$tip.label
wrappedASE(trait = trait, phy = tree, lower_bound = 0, upper_bound = 1)
}
\author{
Mark Juhn
}
