% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedASE.R
\name{wrappedASE}
\alias{wrappedASE}
\title{Ancestral State Estimation in circular space}
\usage{
wrappedASE(trait, phy, lower_bound, upper_bound)
}
\arguments{
\item{trait}{data vector for a single trait, with names matching tips in phy}

\item{phy}{phylogenetic tree of class phylo}

\item{lower_bound}{lower bound for trait in circular space}

\item{upper_bound}{upper bound for trait in circular space}
}
\description{
Estimates ancestral states using wrapped phylogenetic independent contrasts
}
\examples{
tree = ape::rcoal(10)
trait = runif(n = 10, min = 0, max = 1)
names(trait) = tree$tip.label
wrappedASE(trait = trait, phy = tree, lower_bound = 0, upper_bound = 1)
}
\author{
Mark Juhn
}
